+++ src-input/duk_bi_duktape.c
@@ -19,52 +19,10 @@ DUK_INTERNAL duk_ret_t duk_bi_duktape_object_info(duk_context *ctx) {
 }
 
 DUK_INTERNAL duk_ret_t duk_bi_duktape_object_act(duk_context *ctx) {
-	duk_hthread *thr = (duk_hthread *) ctx;
-	duk_activation *act;
-	duk_uint_fast32_t pc;
-	duk_uint_fast32_t line;
 	duk_int_t level;
 
-	/* -1             = top callstack entry, callstack[callstack_top - 1]
-	 * -callstack_top = bottom callstack entry, callstack[0]
-	 */
 	level = duk_to_int(ctx, 0);
-	if (level >= 0 || -level > (duk_int_t) thr->callstack_top) {
-		return 0;
-	}
-	DUK_ASSERT(level >= -((duk_int_t) thr->callstack_top) && level <= -1);
-	act = thr->callstack + thr->callstack_top + level;
-
-	duk_push_object(ctx);
-
-	duk_push_tval(ctx, &act->tv_func);
-
-	/* Relevant PC is just before current one because PC is
-	 * post-incremented.  This should match what error augment
-	 * code does.
-	 */
-	pc = duk_hthread_get_act_prev_pc(thr, act);
-	duk_push_uint(ctx, (duk_uint_t) pc);
-
-#if defined(DUK_USE_PC2LINE)
-	line = duk_hobject_pc2line_query(ctx, -2, pc);
-#else
-	line = 0;
-#endif
-	duk_push_uint(ctx, (duk_uint_t) line);
-
-	/* Providing access to e.g. act->lex_env would be dangerous: these
-	 * internal structures must never be accessible to the application.
-	 * Duktape relies on them having consistent data, and this consistency
-	 * is only asserted for, not checked for.
-	 */
-
-	/* [ level obj func pc line ] */
-
-	/* XXX: version specific array format instead? */
-	duk_xdef_prop_stridx_wec(ctx, -4, DUK_STRIDX_LINE_NUMBER);
-	duk_xdef_prop_stridx_wec(ctx, -3, DUK_STRIDX_PC);
-	duk_xdef_prop_stridx_wec(ctx, -2, DUK_STRIDX_LC_FUNCTION);
+	duk_inspect_callstack_entry(ctx, level);
 	return 1;
 }